bugs and error are fixed

+++ src/controllers/admin.controllers.js
@@ -28,6 +28,8 @@ async function updateUser(req, res) {
     const { id } = req.params
     const { login, email, fullName, password } = req.body
     try {
+        let user = await UserModel.findOne({ _id: id })
+        if (!user) return res.status(200).json({ success: false, message: "User not founded" })
         const salt = await bcrypt.genSalt(12)
         const hashPassword = bcrypt.hash(password, salt)
         await UserModel.findOneAndUpdate({ _id: id }, {
@@ -46,6 +48,8 @@ async function updateUser(req, res) {
 async function deleteUser(req, res) {
     const { id } = req.params
     try {
+        let user = await UserModel.findOne({ _id: id })
+        if (!user) return res.status(200).json({ success: false, message: "User not founded" })
         await UserModel.findOneAndRemove({ _id: id })
         res.json({ success: true, message: "User has been deleted successfully" })
     } catch (error) {
